[add] type difinition file
