+++ src/logics/game/index.ts
@@ -1,8 +1,10 @@
 import { BoardConfig } from '../board';
+import { GameMode } from './types';
 
-export type GameState = 'initialized' | 'playing' | 'completed' | 'failed';
+// types
+export type * from './types';
 
-export type GameMode = 'easy' | 'normal' | 'hard';
+export const GAME_MODE_LIST: GameMode[] = ['easy', 'normal', 'hard'];
 
 export const gameModeToOptions = (gameMode: GameMode): BoardConfig => {
   switch (gameMode) {
@@ -12,7 +14,7 @@ export const gameModeToOptions = (gameMode: GameMode): BoardConfig => {
       return { rows: 16, cols: 16, mines: 40 };
     case 'hard':
       return { rows: 30, cols: 16, mines: 99 };
+    default:
+      return { rows: 9, cols: 9, mines: 10 };
   }
 };
-
-export const GAME_MODE_LIST: GameMode[] = ['easy', 'normal', 'hard'];
