+++ src/logics/game/types.ts
@@ -1,0 +1,3 @@
+export type GameState = 'initialized' | 'playing' | 'completed' | 'failed';
+
+export type GameMode = 'easy' | 'normal' | 'hard';
